hcange float to double on users column

+++ db/migrate/20171129104754_addcolumn_to_user_for_maps.rb
@@ -1,7 +1,7 @@
 class AddcolumnToUserForMaps < ActiveRecord::Migration[5.1]
   def change
     add_column :users, :address, :string
-    add_column :users, :latitude, :float
-    add_column :users, :longitude, :float
+    add_column :users, :latitude, :double
+    add_column :users, :longitude, :double
   end
 end
